split spans into logs and spans
