+++ .gitignore
@@ -11,3 +11,5 @@
 
 .venv
 __pycache__
+
+other/
