+++ pkg/bunapp/chmigrations/20230910231031_initial.down.sql
@@ -12,7 +12,22 @@ DROP TABLE IF EXISTS ?DB.spans_index_buffer ?ON_CLUSTER
 
 DROP TABLE IF EXISTS ?DB.spans_data_buffer ?ON_CLUSTER
 
+--migration:split
 --------------------------------------------------------------------------------
+DROP TABLE IF EXISTS ?DB.logs_index ?ON_CLUSTER
+
+--migration:split
+
+DROP TABLE IF EXISTS ?DB.logs_data ?ON_CLUSTER
+
+--migration:split
+
+DROP TABLE IF EXISTS ?DB.logs_index_buffer ?ON_CLUSTER
+
+--migration:split
+
+DROP TABLE IF EXISTS ?DB.logs_data_buffer ?ON_CLUSTER
+
 --migration:split
 
 DROP VIEW IF EXISTS ?DB.metrics_uptrace_tracing_events_mv ?ON_CLUSTER
@@ -42,7 +57,6 @@ DROP TABLE IF EXISTS ?DB.datapoint_hours ?ON_CLUSTER
 
 DROP VIEW IF EXISTS ?DB.datapoint_hours_mv ?ON_CLUSTER
 
---------------------------------------------------------------------------------
 --migration:split
 
 DROP TABLE IF EXISTS ?DB.service_graph_edges ?ON_CLUSTER
